Changed lat/long data types to decimal?

+++ db/migrate/20120111145221_change_waypoints_float_precision.rb
@@ -1,7 +1,7 @@
 class ChangeWaypointsFloatPrecision < ActiveRecord::Migration
   def up
-    change_column :waypoints, :latitude, :float, :precision => 20, :scale => 16
-    change_column :waypoints, :longitude, :float, :precision => 20, :scale => 16
+    change_column :waypoints, :latitude, :decimal, :precision => 20, :scale => 16
+    change_column :waypoints, :longitude, :decimal, :precision => 20, :scale => 16
   end
 
   def down
